usbserver/requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyusb==1.2.1
Werkzeug==2.3.7
